+++ src/HoneybeeDotNet/Model/EnergyMaterial.cs
@@ -94,10 +94,10 @@ namespace HoneybeeDotNet.Model
         /// <param name="specificHeat">Specific heat of the material layer in J/(kg-K). (required).</param>
         /// <param name="type">type (default to &quot;EnergyMaterial&quot;).</param>
         /// <param name="roughness">roughness (default to RoughnessEnum.MediumRough).</param>
-        /// <param name="thermalAbsorptance">Fraction of incident long wavelength radiation that is absorbed by the material. Default value is 0.9. (default to 0.9M).</param>
-        /// <param name="solarAbsorptance">Fraction of incident solar radiation absorbed by the material. Default value is 0.7. (default to 0.7M).</param>
-        /// <param name="visibleAbsorptance">Fraction of incident visible wavelength radiation absorbed by the material. Default value is 0.7. (default to 0.7M).</param>
-        public EnergyMaterial(string name, decimal thickness, decimal conductivity, decimal density, decimal specificHeat, string type = "EnergyMaterial", RoughnessEnum? roughness = RoughnessEnum.MediumRough, decimal thermalAbsorptance = 0.9M, decimal solarAbsorptance = 0.7M, decimal visibleAbsorptance = 0.7M)
+        /// <param name="thermalAbsorptance">Fraction of incident long wavelength radiation that is absorbed by the material. Default value is 0.9. (default to 0.9).</param>
+        /// <param name="solarAbsorptance">Fraction of incident solar radiation absorbed by the material. Default value is 0.7. (default to 0.7).</param>
+        /// <param name="visibleAbsorptance">Fraction of incident visible wavelength radiation absorbed by the material. Default value is 0.7. (default to 0.7).</param>
+        public EnergyMaterial(string name, double thickness, double conductivity, double density, double specificHeat, string type = "EnergyMaterial", RoughnessEnum? roughness = RoughnessEnum.MediumRough, double thermalAbsorptance = 0.9, double solarAbsorptance = 0.7, double visibleAbsorptance = 0.7)
         {
             // to ensure "name" is required (not null)
             if (name == null)
@@ -170,7 +170,7 @@ namespace HoneybeeDotNet.Model
             // use default value if no "thermalAbsorptance" provided
             if (thermalAbsorptance == null)
             {
-                this.ThermalAbsorptance = 0.9M;
+                this.ThermalAbsorptance = 0.9;
             }
             else
             {
@@ -179,7 +179,7 @@ namespace HoneybeeDotNet.Model
             // use default value if no "solarAbsorptance" provided
             if (solarAbsorptance == null)
             {
-                this.SolarAbsorptance = 0.7M;
+                this.SolarAbsorptance = 0.7;
             }
             else
             {
@@ -188,7 +188,7 @@ namespace HoneybeeDotNet.Model
             // use default value if no "visibleAbsorptance" provided
             if (visibleAbsorptance == null)
             {
-                this.VisibleAbsorptance = 0.7M;
+                this.VisibleAbsorptance = 0.7;
             }
             else
             {
@@ -210,7 +210,7 @@ namespace HoneybeeDotNet.Model
         /// <value>Thickness of the material layer in meters.</value>
         [DataMember(Name="thickness", EmitDefaultValue=false)]
         [JsonProperty("thickness")]
-        public decimal Thickness { get; set; }
+        public double Thickness { get; set; }
 
         /// <summary>
         /// Thermal conductivity of the material layer in W/(m-K).
@@ -218,7 +218,7 @@ namespace HoneybeeDotNet.Model
         /// <value>Thermal conductivity of the material layer in W/(m-K).</value>
         [DataMember(Name="conductivity", EmitDefaultValue=false)]
         [JsonProperty("conductivity")]
-        public decimal Conductivity { get; set; }
+        public double Conductivity { get; set; }
 
         /// <summary>
         /// Density of the material layer in kg/m3.
@@ -226,7 +226,7 @@ namespace HoneybeeDotNet.Model
         /// <value>Density of the material layer in kg/m3.</value>
         [DataMember(Name="density", EmitDefaultValue=false)]
         [JsonProperty("density")]
-        public decimal Density { get; set; }
+        public double Density { get; set; }
 
         /// <summary>
         /// Specific heat of the material layer in J/(kg-K).
@@ -234,7 +234,7 @@ namespace HoneybeeDotNet.Model
         /// <value>Specific heat of the material layer in J/(kg-K).</value>
         [DataMember(Name="specific_heat", EmitDefaultValue=false)]
         [JsonProperty("specific_heat")]
-        public decimal SpecificHeat { get; set; }
+        public double SpecificHeat { get; set; }
 
         /// <summary>
         /// Gets or Sets Type
@@ -250,7 +250,7 @@ namespace HoneybeeDotNet.Model
         /// <value>Fraction of incident long wavelength radiation that is absorbed by the material. Default value is 0.9.</value>
         [DataMember(Name="thermal_absorptance", EmitDefaultValue=false)]
         [JsonProperty("thermal_absorptance")]
-        public decimal ThermalAbsorptance { get; set; }
+        public double ThermalAbsorptance { get; set; }
 
         /// <summary>
         /// Fraction of incident solar radiation absorbed by the material. Default value is 0.7.
@@ -258,7 +258,7 @@ namespace HoneybeeDotNet.Model
         /// <value>Fraction of incident solar radiation absorbed by the material. Default value is 0.7.</value>
         [DataMember(Name="solar_absorptance", EmitDefaultValue=false)]
         [JsonProperty("solar_absorptance")]
-        public decimal SolarAbsorptance { get; set; }
+        public double SolarAbsorptance { get; set; }
 
         /// <summary>
         /// Fraction of incident visible wavelength radiation absorbed by the material. Default value is 0.7.
@@ -266,7 +266,7 @@ namespace HoneybeeDotNet.Model
         /// <value>Fraction of incident visible wavelength radiation absorbed by the material. Default value is 0.7.</value>
         [DataMember(Name="visible_absorptance", EmitDefaultValue=false)]
         [JsonProperty("visible_absorptance")]
-        public decimal VisibleAbsorptance { get; set; }
+        public double VisibleAbsorptance { get; set; }
 
         /// <summary>
         /// Returns the string presentation of the object
@@ -434,8 +434,8 @@ namespace HoneybeeDotNet.Model
                 yield return new System.ComponentModel.DataAnnotations.ValidationResult("Invalid value for Name, length must be greater than 1.", new [] { "Name" });
             }
 
-            // Thickness (decimal) maximum
-            if(this.Thickness > (decimal)3)
+            // Thickness (double) maximum
+            if(this.Thickness > (double)3)
             {
                 yield return new System.ComponentModel.DataAnnotations.ValidationResult("Invalid value for Thickness, must be a value less than or equal to 3.", new [] { "Thickness" });
             }
@@ -447,32 +447,32 @@ namespace HoneybeeDotNet.Model
                 yield return new System.ComponentModel.DataAnnotations.ValidationResult("Invalid value for Type, must match a pattern of " + regexType, new [] { "Type" });
             }
 
-            // ThermalAbsorptance (decimal) maximum
-            if(this.ThermalAbsorptance > (decimal)0.99999)
+            // ThermalAbsorptance (double) maximum
+            if(this.ThermalAbsorptance > (double)0.99999)
             {
                 yield return new System.ComponentModel.DataAnnotations.ValidationResult("Invalid value for ThermalAbsorptance, must be a value less than or equal to 0.99999.", new [] { "ThermalAbsorptance" });
             }
 
-            // SolarAbsorptance (decimal) maximum
-            if(this.SolarAbsorptance > (decimal)1)
+            // SolarAbsorptance (double) maximum
+            if(this.SolarAbsorptance > (double)1)
             {
                 yield return new System.ComponentModel.DataAnnotations.ValidationResult("Invalid value for SolarAbsorptance, must be a value less than or equal to 1.", new [] { "SolarAbsorptance" });
             }
 
-            // SolarAbsorptance (decimal) minimum
-            if(this.SolarAbsorptance < (decimal)0)
+            // SolarAbsorptance (double) minimum
+            if(this.SolarAbsorptance < (double)0)
             {
                 yield return new System.ComponentModel.DataAnnotations.ValidationResult("Invalid value for SolarAbsorptance, must be a value greater than or equal to 0.", new [] { "SolarAbsorptance" });
             }
 
-            // VisibleAbsorptance (decimal) maximum
-            if(this.VisibleAbsorptance > (decimal)1)
+            // VisibleAbsorptance (double) maximum
+            if(this.VisibleAbsorptance > (double)1)
             {
                 yield return new System.ComponentModel.DataAnnotations.ValidationResult("Invalid value for VisibleAbsorptance, must be a value less than or equal to 1.", new [] { "VisibleAbsorptance" });
             }
 
-            // VisibleAbsorptance (decimal) minimum
-            if(this.VisibleAbsorptance < (decimal)0)
+            // VisibleAbsorptance (double) minimum
+            if(this.VisibleAbsorptance < (double)0)
             {
                 yield return new System.ComponentModel.DataAnnotations.ValidationResult("Invalid value for VisibleAbsorptance, must be a value greater than or equal to 0.", new [] { "VisibleAbsorptance" });
             }
